requirements.txt
toml
sqlalchemy==1.3.20
google-cloud-speech
google-cloud-storage
python-telegram-bot==13.4
tinytag==1.5.0
alembic==1.4.3
pymediainfo==5.0.3
